Update rtn requirements to new unique reference
https://eaflood.atlassian.net/browse/WATER-5085

When testing the results of our one-off [quarterly return versions for water companies job](https://github.com/DEFRA/water-abstraction-system/pull/1648), the users spotted an issue with the return references being generated.

The way we create return requirement references differs from NALD's approach. NALD now uses unique references, whereas we don't. We assumed the ID was based on how the rest of the IDs are generated in NALD: they are unique to the region the record is linked to, but you can and will get duplicates across the regions.

Great, we can ditch our logic and just make `legacy_id` an incrementing ID field, can't we?

Unfortunately, no. The problem we've got is that it appears NALD only began creating unique references from approximately 2008 onwards. So, about 7K  duplicate references exist that (at this time) we can't just go in and change.

Therefore, [we've made a change](https://github.com/DEFRA/water-abstraction-service/pull/2695) that resolves this issue. It adds a new column called `reference` to `water.return_requirements`, which will serve as an incrementing ID. As new records are added in the future, we can leave PostgreSQL to determine what the next reference should be. We initialise the column to be the existing legacy ID where it is unique. For the duplicates, one return requirement per duplicate group gets to keep the legacy ID. For the rest we assign new references. We finish it off by adding a new trigger that will fire when a new `return_requirement` is added. It will copy the newly created `reference` to `legacy_id` to avoid breaking any existing functionality. Though, if it's going to be of use in the future, we'll want this to become the 'actual' reference and deprecate the `legacy_id` field.

That sorts the data side. We now need to update our logic.

- We need to drop anything that is currently trying to assign a legacy_id
- We need to update our models and helpers to be aware of the new field
- We need to switch to using `reference` instead of `legacy_id` when it comes to generating return logs