Update the return requirements helper
We drop setting the legacy ID as this will be handled by the tigger we've added to the table, and reference is generated automatically.

We also stop populating externalID. We shouldn't have been doing this as it is intended to identify the NALD record, which now wouldn't exist.

There are times in tests however where sometimes we need these values to be populated. For example, we are using a POJO to stand in for an actual record. So, we don't drop `generateLegacyId()` but we do rename it to `generateReference()`.