Fix broken link from licence to billing account (#2058)
https://eaflood.atlassian.net/browse/WATER-5090

**Context**

We have been working on migrating the legacy view billing account page to a new page (see WATER-4988 ).

That work is behind a feature flag, so it is not visible to users at this time. In our version of the page, we need the link on the Bills tab in 'view licence summary' to pass both the billing account ID and licence ID. In the legacy version, it's just the billing account ID.

The feature flag determines the kind of link we use.

**Problem**

If you click that link on the Bills tab, you'll see the ”Sorry, there is a problem with the service” message.

It is because we've made a mistake in our code. We're determining the 'root path' for the link using the feature flag. However, we're adding `?licence-id=170b97b1-d65f-48ef-8807-f3cf5e97687b` to both the new and legacy versions. That should also be something determined by the feature flag.

**Fix**

Add the licence ID where we also determine which link to use, instead of being hard-coded into the view.