Handle empty abstraction period in return log
https://eaflood.atlassian.net/browse/WATER-5051

It was spotted that we have (currently) 2,271 return requirements with no abstraction period out of total 85,567.

If the annual return log creation job encounters one it throws an error.

Also, the NALD generated return logs based on these return requirements are missing the abstraction period information in their metadata field, which breaks our new view return log page.

Because their are further issues caused by this missing data when it comes to reporting, the long term intent is to fix the data itself. Till then, we need to ensure the service doesn't break in the short term.