Handle empty abstraction period in return log (#2062)
https://eaflood.atlassian.net/browse/WATER-5051

It was spotted that we have (currently) 2,271 return requirements with no abstraction period out of a total of 85,567.

If the annual return log creation job encounters one, it throws an error.

Additionally, the NALD-generated return logs, based on these return requirements, are missing the abstraction period information in their metadata field, which breaks our new view return log page.

Because further issues arise from the missing data when it comes to reporting, the long-term goal is to correct the data itself. In the meantime, we need to ensure the service doesn't break in the short term.