Fix select financial year for 2PT supplementary (#2064)
https://eaflood.atlassian.net/browse/WATER-5091

In [Fix invalid fin year in 2PT annual bill run setup](https://github.com/DEFRA/water-abstraction-system/pull/2056), we made a change to the bill run `YearValidator` to fix an issue with the annual two-part tariff bill run journey.

However, we were only considering the annual journey when we made the change. We didn't stop to think that the page is also used in the two-part tariff supplementary journey.

Doh!

This change fixes the validator to handle both journeys.