Fix duplicate points in return versions setup
https://eaflood.atlassian.net/browse/WATER-5069

During UAT of the manual return version set up journey it was spotted that duplicate points can be listed in the "Select the points for the requirements for returns" screen.

The reason is a licence that has more than one purpose linked to the same point. The current fetch is compiling a list of every point linked to each purpose against the licence's current version. But if Point A is linked to purpose 1 and 2, it results in Point A being listed twice.

This change updates the `FetchPointsService` to ensure a point is only added once to the list it compiles.

We also found that sometimes the unit test would fail because the order the points were returned was not consistent, yet the tests assumed they would be. So, we also order the points by their `ngr1` property, which we know will always be populated.