Fix duplicate points in return versions setup (#2066)
https://eaflood.atlassian.net/browse/WATER-5069

During the UAT of the manual return version setup journey, it was noted that duplicate points can be listed on the "Select the points for the requirements for returns" screen.

The reason is a licence that has more than one purpose linked to the same point. The current fetch is compiling a list of every point linked to each purpose against the licence's current version. But if Point A is linked to purposes 1 and 2, it results in Point A being listed twice.

This change updates the `FetchPointsService` to ensure a point is only added once to the list it compiles.

We also found that sometimes the unit test would fail because the order in which the points were returned was inconsistent, yet the tests assumed it would be consistent. So, we also order the points by their `ngr1` property, which we know will always be populated.