Fix notices reminders returns period page title (#2069)
We have accidentally hard codded the page title to 'invitations'. When we are on the reminders page we should show 'Select the returns periods for the reminders'

This change updates the page title to use the session to show the correct text.

This change also lift the page title out of the services and into the presenter. This reduces duplicate logic.