Add source to abstraction alerts notifications (#2054)
* Add source to abstraction alerts notifications

https://eaflood.atlassian.net/browse/WATER-5067

We have previously implemented part of the notification presenter. We left out some of the personalisation to be done in later changes.

This is one of those changes.

This change adds the 'source' to the personalisation. This is conditional logic when the monitoring station has a river name.

If it has a name we will show the river name with some additional text.

If there is no river name it will be set in the database as either null of an empty string (''). When this is the case we want to send notify an empty string so it does not render anything in the notification.