Add condition text to abstraction alerts notifications (#2068)
* Add condition text to abstraction alerts notifications

https://eaflood.atlassian.net/browse/WATER-5067

We have previously implemented part of the notification presenter. We left out some of the personalisation to be done in later changes.

This is one of those changes.

This change adds the 'condition_text' to the personalisation. This is conditional logic when the monitoring station has a licence version purpose and that purpose has notes.

The notes could be a string, an empty sting or null.