Amend "licence hands-off flow or level threshold" page to use radios (#2038)
We spotted that the "What is the licence hands-off flow or level threshold?" page (the first one encountered during the monitoring stations "Tag a licence" journey) uses a drop-down select component. The GOV.UK Design System recommends using radio buttons for this type of question. We therefore update the page to use these.