Monitoring stations - Remove tag button https://eaflood.atlassian.net/browse/WATER-5032
In PR https://github.com/DEFRA/water-abstraction-system/pull/2020 the remove tag functionality has been moved to a new page which displays the linked monitoring stations for a licence.

Following this change the "Remove tag" button can now be removed from the View monitoring station page.

The "Tag a licence"  button is also going to be moved up the page to make it more prominent.