Add templateId to abstraction alerts notifications
https://eaflood.atlassian.net/browse/WATER-5067

We have previously implemented part of the notification presenter. We left out some of the personalisation to be done in later changes.

This is one of those changes.

This change adds the 'templateId' logic to the abstraction alerts notifications.

The logic is similar to the message ref. But they have been kept separate for readability.

But we could look to refactor in another change.