Update fetch abstraction alert recipients service
https://eaflood.atlassian.net/browse/WATER-5067

We have previously implemented logic around getting recipients. This was wen a primary user was set then it would have priority over the licence holder, also known as a registered licence taking priority over an unregistered licence.

If either of these had an additional contact then we would also include them.

This is no longer the case.

This change update the existing logic to prioritise the additional contact over the primary user and licence holder.

So when one (or more) additional contacts are found for a licence this recipient/s will receive the notice.

If there are no additional contacts then we fall into our normal registered vs unregistered flow.