Monitoring stations - Remove tag button (#2071)
https://eaflood.atlassian.net/browse/WATER-5032

In PR https://github.com/DEFRA/water-abstraction-system/pull/2020 the remove tag functionality has been moved to a new page which displays the linked monitoring stations for a licence.

Following this change the "Remove tag" button can now be removed from the View monitoring station page. However, if the `ENABLE_LICENCE_MONITORING_STATIONS_VIEW` feature flag is `false` the "Remove tag" button will still be visable. This is because the new functionality to remove the tag is hidden behind this feature flag so woul be inaccessible.

The "Tag a licence"  button is also going to be moved up the page to make it more prominent.