Add templateId to abstraction alerts notifications (#2074)
* Add messageRef to abstraction alerts notifications

https://eaflood.atlassian.net/browse/WATER-5067

We have previously implemented part of the notification presenter. We left out some of the personalisation to be done in later changes.

This is one of those changes.

This change adds the 'messageRef' logic to the abstraction alerts notifications. This is saved in the database and used on the notices manage page.

This logic will be duplicated in another change to assign the correct Noitify template id.

* chore: add default message

* Add templateId to abstraction alerts notifications

https://eaflood.atlassian.net/browse/WATER-5067

We have previously implemented part of the notification presenter. We left out some of the personalisation to be done in later changes.

This is one of those changes.

This change adds the 'templateId' logic to the abstraction alerts notifications.

The logic is similar to the message ref. But they have been kept separate for readability.

But we could look to refactor in another change.