Update abstraction alerts create notice and confirmation (#2078)
https://eaflood.atlassian.net/browse/WATER-5097

The abstraction alerts journey is close to completion from end to end.

When we create a notice we need to set the licences for the legacy notices page to render.

This is not used but breaks if it is not set.

This change updates the create notice presenter to include the recipients licence refs.

There is also an issue showing undefined in the confirmation banner. The page title has been updated to reflect the water abstraction alerts.