Refactor session cleanup job to be general clean
https://eaflood.atlassian.net/browse/WATER-5075

With WRLS taking over management of returns, plus the new quarterly returns functionality, there is likely to be an increase in the number of VOID returns generated.

In NALD, the return logs were only generated at the point where the business would send out invites to customers to submit them. In WRLS, we are generating for the current cycle, ahead of when we send the invite. Should a user need to make a change, for example, due to an error in NALD, nothing would need to be voided. In WRLS, at least one return log will need to be. If it's a licence with a quarterly requirement, it could be four.

Knowing this, we don't want the voids to become 'noise' and start swamping the actual return logs within the service.

So, we intend to add a job we can trigger that will delete all empty (not submitted or received) return logs with a status of `void`, very much like we are already doing for expired temporary sessions. Neither does anything complex, plus we'll also add an extra process for cleaning empty bill runs (we've been inspired!)

It therefore makes sense to just trigger a single 'clean' job that triggers all the cleaning tasks. The first step is to refactor the existing session clean up ready for adding the additional steps.