Add clean empty bill runs task to clean job
https://eaflood.atlassian.net/browse/WATER-5075

We already had an expired sessions cleanup job that we [refactored](#2080) to allow us to add other 'cleaning' tasks within a single 'clean' job.

This was because of a need to manage the number of 'VOID' returns we are likely to be generating what with the way WRLS manages returns (creating them ahead of time rather than at point of submission).

The inspiration for simply deleting 'void' return logs rather than keeping them around was 'empty' bill runs.

When a user creates a bill run but the engine determines there is nothing to bill, the service doesn't just flash a message. The bill run engine relies on a record to exist, so instead it flags the bill run record as 'empty'.

The bill run is still listed with its status displayed as 'EMPTY', just like a 'VOID' return log. And we allows users to delete them rather than having them cluttering up the results and degrading the performance of the service.

So, we asked the business, why not just delete them? They agreed hence [Add clean empty void return logs task to clean job](#2081).

As a final step, and to keep things consistent, this adds a clean empty bill runs task, so the service can pick up the ones they've missed, or that are hidden away in the history!