Show warning when view VOID return logs
https://eaflood.atlassian.net/browse/WATER-5076

With WRLS taking over management of returns, we updated the view a return log functionality. We now display the details of a VOID return log, a feature the legacy service was unable to provide.

However, there is concern that users may overlook the fact that the return log has been voided and use this information inaccurately.

So, they have asked us to display a warning message on the view return log page when the status is `void`.