Split up existing return logs fixture
We intended to add a return-logs fixture than found there was already one. However, there is a lot going on in there and adding our stuff would just overwhelm it.

So, this is a bit of refactoring to just split the various parts into their own fixtures to make way for what we want to add.