Add clean empty void return logs task to clean job (#2081)
https://eaflood.atlassian.net/browse/WATER-5075

With WRLS taking over management of returns, plus the new quarterly returns functionality, there is likely to be an increase in the number of VOID returns generated.

In NALD, the return logs were only generated at the point where the business would send out invites to customers to submit them. In WRLS, we are generating for the current cycle, ahead of when we send the invite. Should a user need to make a change, for example, due to an error in NALD, nothing would need to be voided. In WRLS, at least one return log will need to be. If it's a licence with a quarterly requirement, it could be four.

Knowing this, we don't want the voids to become 'noise' and start swamping the actual return logs within the service.

In [Refactor session cleanup job to be general clean](https://github.com/DEFRA/water-abstraction-system/pull/2080), we refactored an existing job to enable the addition of further cleaning tasks.

This change adds the key task that initiated these changes: deleting 'empty' `VOID` return logs.

By empty, we mean

- The return log has not been marked as received before it was made `void`
- The return log has no submissions against, including 'nil returns'