Reinstate backlink property
This was one of the tests which appeared to have gotten very confused about what is going on.

For a start, return logs don't have versions, it is the return submissions that do. So, there is no such thing as the 'latest' return log.

Second, the test was instantiating ReturnVersion version instances to populate in the `returnLog.versions` property. Return versions have _nothing_ to do with `returnLog.versions`, and nothing in the code fetches and uses these things. We guess the fact the property is called `versions` and we have return versions is what confused things.

It just so happens that return versions have the same properties that the presenter needs to extract, hence the test passes.