Drop meter details property test
Similar to `licenceRef`, we wouldn't add a test for this kind of property either.

Unlike `licenceRef`, it isn't just returning a value on the object passed in. But it is just calling a function that is already tested in `base-return-logs.presenter.test.js`. Again, the `it('correctly presents the data', ` test provides sufficient coverage.