Reinstate the table title property test
Reinstating this test exposed that there were some things we were doing in the presenter that could be improved.

For example, it seemed misleading to return a title when we wouldn't be showing the table, so we updated the presenter to pass in the summary table data to the function so it could use this to determine whether a title was needed or not.

All 3 of our new tests passed originally, until we tweaked one to be a nil return to spice things up! Good job we did because it failed.

It had exposed that we were still calling `generateSummaryTableHeaders()` and `generateSummaryTableRows()` even though we were dealing with a nil return, so no table will be shown.

Another tweak to the presenter and all is well again.