Fix no return submissions data setup
There was a couple of lines in the presenter that test coverage said we were never hitting. When we dug in, we found it was because of an incorrect assumption when it came to how the `FetchReturnLogService` was working.

We assumed if there were no return submissions, it would return a return log instance with an empty array for return submissions. Instead, it doesn't even bother to query for them, so the property is never set.

Therefore, where we are representing a return log with no submission data, it is more accurate to delete the property rather than set it to a blank array.

And we finally got those missing lines covered!