Update and correct view return log presenter test  (#2086)
https://eaflood.atlassian.net/browse/WATER-5076

We started work on [adding a warning message to the view return log page](https://github.com/DEFRA/water-abstraction-system/pull/2083), which is a relatively simple change.

However, when we came to update the tests for the presenter, we found

- Tests that weren't testing the thing they thought they were
- Tests generating the wrong instance types, but passing because of similar properties
- Tests generating instances we don't need
- Inconsistent structure with other presenter tests (no `it('correctly presents the data')` test)
- Need to use proxyquire to stub out presenters (we don't normally stub out presenters in tests)

So, being the good 'boy scouts' we are, this change is about sorting these issues with the existing tests before we then make further changes to them.