Use relevant not current lic ver in rtn ver setup
https://eaflood.atlassian.net/browse/WATER-5096

After switching to WRLS to manage everything to do with returns, including the setup of new return versions the users have spotted an issue.

When we built the return version setup, the assumption was that the current licence version should be used for all journeys (start with abstraction data, copy existing, and manual).

But users have encountered an error when using 'copy existing' when copying a return version created using an older licence version.

The logic is trying to copy return requirements points from the previous return version, and match them to the ones in the current licence version. If however the points have changed and they no longer exist, the journey blows up.

What it has exposed is that we got it wrong when it comes to which licence version to base our data on. They have told us we should be using the data from the licence version that matches the start date selected.

For example, image the following scenario. These are the current return versions.

| ID | Start date | End date   | Status   |
|----|------------|------------|----------|
| R2 | 2022-04-01 |            | Approved |
| R1 | 2019-05-13 | 2022-03-31 | Approved |

And these are the licence versions.

| ID | Start date | End date   | Current |
|----|------------|------------|---------|
| V2 | 2022-04-16 |            | Yes     |
| V1 | 2019-05-13 | 2022-04-15 | No      |

> We've purposefully made the dates not line up to demonstrate this is often the case.

Between V1 and V2 the points were changed on the purposes. But at the moment the points on R1 and R2 will be the same, and both based on V1, the defunct version.

**Where we error**

A user goes to add a new return version starting 2025-04-01. They opt to copy the previous return version R2. We error because we cannot find the points assigned to R2, from all those linked to V2, because we're taking all our lookup/validation data _for all journeys_ from V2.

**Where we are wrong**

A user goes to add a new return version starting 2022-04-01 (they are going to replace the R2). They opt for either the 'start from abstraction data', or manual journey routes (it does not matter). The return version will be wrong because it is using V2 for all data, even though it did not apply at that point in time.

**The fix**

We need to be selecting the licence version we take our data from based on the 'start date' entered by the user. Not only will this fix the error they are seeing, it will ensure the return version they setup is 'correct'.