Add clean empty bill runs task to clean job (#2082)
https://eaflood.atlassian.net/browse/WATER-5075

We already had an expired sessions cleanup job that we [refactored](#2080) to allow us to add other 'cleaning' tasks within a single 'clean' job.

This was due to a need to manage the number of 'VOID' returns we are likely to generate, given the way WRLS manages returns (by creating them ahead of time rather than at the point of submission).

The inspiration for simply deleting 'void' return logs rather than keeping them around was 'empty' bill runs.

When a user creates a bill run but the engine determines there is nothing to bill, the service doesn't just flash a message. The bill run engine relies on a record to exist; therefore, it flags the bill run record as 'empty'.

The bill run is still listed, with its status displayed as 'EMPTY', similar to a 'VOID' return log. We allow users to delete them, rather than having them clutter the results and degrade the service's performance.

So, we asked the business, Why not just delete them? They agreed, hence [Add clean empty void return logs task to clean job](#2081).

As a final step, and to maintain consistency, this adds a clean empty bill runs task, allowing the service to pick up the ones they've missed or that are hidden in the history!