Update FetchAbstractionData to find relevant ver
This is the first step in updating the 'Start with abstraction' journey.

We update the fetch service to find the relevant licence version, rather than just the current one. Because of the complexity, we felt it would be easier to locate any relevant licence agreement in a separate query, rather than as a raw subquery.