Update FetchPurposes to use relevant licence ver
This is the update to FetchPurposesService to get it to find the relevant licence version in order to pick purposes to show to the user.

However, this change has made us realise this is not the way to go. We now need to update the unit tests for this service to ensure they cover this complexity. We realised if we continue with this approach, updating each service with the logic needed, not only do they risk looking as complex as this, but it means lots of work and duplication when it comes to the testing.

So, we are going in a different direction!