Update notices abstraction alerts status update (#2089)
https://eaflood.atlassian.net/browse/WATER-5100

We have a requirement to update the licence monitoring stations 'Last type of alert sent and date issued'. To do this and keep the exising notification code / logic generic we need to add the licence monitoring station id and the users chosen alert type into the personalisation.

This will not be shown in the template. We initially built the notices functionality around the legacy code. This mean we store the personalisation data as json for each notice we send.

This change adds the licence monitoring station id and the users chosen alert type into the personalisation and saves it in the database.

We need to make another change (in a separate code change) to the notification job, which currently updates the status from notify, to also check if the notification is a water abstraction alert. And if so we need to update the licence monitoring stations 'status_updated_at' and 'status' fields, with the data added by this change.