Add new FetchRelevantLicenceVersionService
We're moving all the logic to find the relevant licence version into its own service, which we can then write dedicated tests for.

We'll call this service when the start date is submitted and save its results to the session. We can then simply extract the licence version ID from the session in later pages and pass it to the various fetch services. This makes their work much simpler (and easier to test for us).