Tidy up unneeded `no_conditions` string
When we first started developing the feature, the ticket stated that if there are no conditions on the licence, we display a single radio button with a suitable option -- we used `no_conditions` to indicate this. This was later changed to the current approach of displaying some static text with a Continue button. We therefore remove the unneeded `no_conditions` string (as the equivalent option when there are conditions uses `not_listed`)