Add new DetermineTwoPartTariffAgreementService
The first thing we have to go back and update is the changes we made to the 'Start with abstraction data' journey.

We updated that to work out the relevant licence version. But it also needs to work out whether a two-part tariff licence agreement was in place on the start date selected.

We updated it to use a similar logic to finding the relevant licence version. But we've since realised we iot it wrong, plus it still needs access to the start date value.

So, it makes better sense to move it into its own service, so we can test and document it in isolation, and call it with the values it needs at the appropriate time.