Update FetchAbstractionDataService
Now we can pass the specific licence version ID we want it to fetch and grab abstraction data from. We can also drop everything to do with licence agreements as they are covered by our new DetermineTwoPartTariffAgreementService.

Because licence agreements are handled outside the service, we also don't need the seeder to generate any.