Fix abstraction alerts multiple licence refs
https://eaflood.atlassian.net/browse/WATER-5067

When we get the 'recipients', we receive the associated licence refs in a comma separate string e.g 'AB12,BC,56'.

When this occurred we were not handling this a comma seperated string but as a licence ref. The download logic had already solved this problem but the implementation was missed in the notices.

This change updates the abstraction alert presenter to check if the licence monitoring stations licence ref is in the recipients  licence refs.