Add new DetermineRelevantLicenceVersionService
We've sorted the 'Start with abstraction data' and 'Manual' journeys (you haven't seen the change to `SubmitStartDate` yet but we're getting there!)

The final journey to tackle is 'Copy from existing', and this was the one that blew up and highlighted we'd gotten it wrong by only using the current licence version.

Unlike the other journeys, this one has some extra complexity. For them, as long as you have the right licence version, we build up the requirements from its abstraction data. The only difference is whether that is done automatically or manually.

Copy existing before this change was relatively simple: just list out all existing return versions and copy the selected one's data into the new return version.

You can't do that, if the abstraction data the existing return version is built from does not match the data on the relevant licence version.

So, now once a user has selected their start date, we not only need to identify the relevant licence version for it, but also which of the existing return versions were built with data from it. Only these are 'copyable candidates' for the 'Copy from existing' journey.

That's a bunch of work we have to do at the point the user selects a start date, that was too much to cram into `SubmitStartDateService`.

Hence, we have created DetermineRelevantLicenceVersionService, which the submit service will call and whose result it will persist into the session for later services to use.