reorder which return version end date to use (#2095)
https://eaflood.atlassian.net/browse/WATER-5099

This issue was found in production when users created new return versions for water companies that need quarterly returns enabled. When inserting a new return version between other existing return version it is setting the end date of one of the return versions incorrectly. This was due to the the order of the functions that determine which end date to use. This fix reorders the function calls with the more specific one higher in the order.