Update Submit Start Date service
The final change we need to make. When the start date is submitted is when we have to make our decisions about the rest of the journey.

When first visiting it is about determining the relevant licence version, what return versions are copyable based on that result, and saving the result to the session.

When re-visiting, we have to determine whether the user has made a change to the start date that requires us to reset the session because the relevant licence version has changed, or it remains the same so they can keep what they have previously submitted.