Update backlink for billing account page via charge information
https://eaflood.atlassian.net/browse/WATER-4988

While testing fixes to the view billing account page, it was noticed there is a link to the billing account page via the via charge information page. This
is part of the legacy and has been [updated to redirect to the new billing account page](https://github.com/DEFRA/water-abstraction-ui/pull/2706). As a result, when users reach the billing account page through the charge
information page, the backlink should take them back there.

This PR updates the backlink on the view billing account page to redirect to the charge information page.