Use relevant not current lic ver in rtn ver setup (#2088)
https://eaflood.atlassian.net/browse/WATER-5096

After switching to WRLS to manage everything related to returns, including the setup of new return versions, users have identified an issue.

When we built the return version setup, the assumption was that the current licence version should be used for all journeys (start with abstraction data, copy existing, and manual).

But users have encountered an error when using 'copy existing' when copying a return version created using an older licence version.

The logic is trying to copy return requirements points from the previous return version, and match them to the ones in the current licence version. If, however, the points have changed and they no longer exist, the journey blows up.

What it has exposed is that we got it wrong when it comes to which version of the licence to base our data on. They have informed us that we should be using the data from the licence version that matches the selected start date.

For example, imagine the following scenario. These are the current return versions.

| ID | Start date | End date   | Status   |
|----|------------|------------|----------|
| R2 | 2022-04-01 |            | Approved |
| R1 | 2019-05-13 | 2022-03-31 | Approved |

And these are the licence versions.

| ID | Start date | End date   | Current |
|----|------------|------------|---------|
| V2 | 2022-04-16 |            | Yes     |
| V1 | 2019-05-13 | 2022-04-15 | No      |

> We've purposefully made the dates not line up to demonstrate this is often the case.

Between V1 and V2, the points were changed for various purposes. But the points on R1 and R2 will be the same, as both are based on V1, the defunct version.

**Where we error**

A user goes to add a new return version starting 2025-04-01. They opt to copy the previous return version R2. We encounter an error because we cannot find the points assigned to R2 among all those linked to V2, as we're retrieving all our lookup/validation data for _all_ journeys from V2.

**Where we are wrong**

A user attempts to add a new return version starting on 2022-04-01 (they plan to replace R2). They opt for either the 'start from abstraction data' or manual journey routes (it does not matter). The return version will be incorrect because it uses V2 for all data, even though it was not applicable at that point in time.

**The fix**

We need to select the licence version we use for our data based on the 'start date' entered by the user. Not only will this fix the error they are seeing, it will ensure the return version they set up is 'correct'.

---

As a result, there are additional changes to how the journey now operates.

We've explained how the abstraction data used in the "start with abstraction data" and "manual" journeys will now be determined by the relevant licence version for the start date selected by the user.

But this also impacts which return versions are copyable, if any. We can only allow users to copy return versions that align with the relevant licence version determined by the start date. Else it will be no different than when we just used the 'current' licence version for everything.

The other change is if a user chooses to amend the start date. If they do this, we have to re-determine the relevant licence version. If we obtain a different result, we must reset the session, as the choices made now are based on redundant abstraction data.