Update notices abstraction alerts status update (#2094)
* Update notices abstraction alerts status update

https://eaflood.atlassian.net/browse/WATER-5100

We have a requirement to update the licence monitoring stations 'Last type of alert sent and date issued'. To do this and keep the exising notification code / logic generic we need to add the licence monitoring station id and the users chosen alert type into the personalisation.

This will not be shown in the template. We initially built the notices functionality around the legacy code. This mean we store the personalisation data as json for each notice we send.

This change updates the notification job to check if the notifications id for an abstraction alert. When it is we update licence monitoring stations 'status_updated_at' and 'status' fields.

This is known as the 'Last alert sent'