Fix permissions for creating & viewing return ver.
https://eaflood.atlassian.net/browse/WATER-5104

Don't ask me how we've ended up here, but we have the permissions the wrong way round!

Viewing a return is scoped to only those with the `billing` role, whereas we never put any auth on the buttons. We think it is because to access the 'Licence setup' tab you need the role `view_charge_versions``, and we assumed that if you can see the tab, it's because you need to administer stuff on it.

However, dipping back into the legacy version, and our own code I can see that they differentiate between viewing 'licence setup' (Charge information on the legacy page) and editing stuff in it.

We need to check the roles of the user, specifically whether they have `billing` as part of determining whether the return version setup buttons are shown.

WE also need to change the scope on viewing a return version to be `view_charge_versions`. This aligns with being able to access the licence setup tab and viewing charge versions.