Fix instantaneous abstraction amount measure text
https://eaflood.atlassian.net/browse/WATER-5068

The users have highlighted the measure we display next to the instantaneous abstraction amount in the view licence purposes page is wrong.

We say '9.80 cubic metres per second' when it should be '9.80 litres per second'.

As part of this we've also spotted that if a licence has multiple purposes, and therefore multiple abstraction amounts, the link we display on the summary tab next to **Abstraction amounts** takes you to the legacy view licence purposes page.

So, in this change we also update the link to take you to our, now fixed, purposes page.