Fix instantaneous abstraction amount measure text (#2103)
https://eaflood.atlassian.net/browse/WATER-5068

The users have highlighted that the measure we display next to the instantaneous abstraction amount in the view licence purposes page is wrong.

We say '9.80 cubic metres per second' when it should be '9.80 litres per second'.

As part of this, we've also noticed that if a licence has multiple purposes, and therefore multiple abstraction amounts, the link displayed on the summary tab next to **Abstraction amounts** directs you to the legacy view licence purposes page.

In this update, we also change the link to direct you to our now fixed purposes page.