Return generated return log ids (#2104)
https://eaflood.atlassian.net/browse/WATER-5099

As part of the testing of this ticket QA have found that when inserting a new return version before an existing one that already has return logs created for it those return logs were incorrectly being voided.

This was due to the code around the on conflict not working as expected. This PR removes that logic and just returns all of the id's that were attempted to be inserted so that any other return logs will be correctly voided.