Fix missed wrong abstraction amount measure text
https://eaflood.atlassian.net/browse/WATER-5068

In [Fix instantaneous abstraction amount measure text](https://github.com/DEFRA/water-abstraction-system/pull/2103) we dealt with an issue where we were displaying the wrong text. We were showing 'cubic metres' when it should have been litres.

Or so we thought!

We missed where a licence has just one purpose with an instantaneous abstraction amount. This means on the licence summary tab, we don't show the text 'Multiple abstractions', instead we just show the amounts, and we're using the wrong measure.

This fixes that missed scenario.